Changed Item to abstract class (we only ever use its children classes for objects)
Added ArmorSlot enum and WeaponEquip enum to designate equipment slots they go in or whether a weapon is two handed.

built removeitem method in InventoryManager. Method takes and item INDEX as input and removes the item in that index of the inventoryList.

Added variables for inventory size and weight constraints in inventory class